[Feat] 🍺 예약내역 헤더 통합
