+++ src/components/Layout/ReservationLayout/index.tsx
@@ -1,13 +1,19 @@
 import ReservationHeader from '@/components/Header/ReservationHeader';
 import Nav from '@/components/Nav/Nav';
-import Tabs from '@/components/Tabs/Tabs';
+// import Tabs from '@/components/Tabs/Tabs';
 import { Outlet, useLocation } from 'react-router-dom';
 
 export function ReservationLayout() {
-  const noLayoutPaths = ['/reservation'];
   const location = useLocation();
+  const noLayoutPaths = [
+    '/reservation/visit/',
+    '/reservation/call/',
+    '/reservation/inquiry/,',
+  ];
 
-  const hasNoLayout = noLayoutPaths.some((path) => location.pathname === path);
+  const hasNoLayout = noLayoutPaths.some((path) =>
+    location.pathname.startsWith(path)
+  );
   if (hasNoLayout) {
     return <Outlet />;
   }
@@ -15,13 +21,7 @@ export function ReservationLayout() {
   return (
     <>
       <ReservationHeader />
-      <div className='w-[90%] justify-self-center pt-[4rem]'>
-        <Tabs
-          leftValue='visit'
-          leftName='방문 상담'
-          rightValue='call'
-          rightName='전화 / 1:1 상담'
-        />
+      <div className='h-screen w-[90%] justify-self-center pt-[12.5rem]'>
         <Outlet />
       </div>
       <Nav />
