+++ src/pages/Reservation/Call/index.tsx
@@ -1,24 +1,9 @@
 import '@/index.css';
-import { useState } from 'react';
 import CallList from '@/components/Reservation/CallList';
-import { ChangeToggle } from '@/components/Reservation/ChangeToggle';
 import { callListData } from '@/mock/mockReservationCall';
 import { ReactComponent as RefreshIcon } from '@/assets/icons/reservation/refresh.svg';
 
 export function ReservationCallPage() {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const [selectedTab, setSelectedTab] = useState('전화 상담 내역');
-
-  const toggleCallList = () => {
-    setIsOpen(!isOpen);
-  };
-
-  const selectTab = (tabName: string) => {
-    setSelectedTab(tabName);
-    setIsOpen(false);
-  };
-
   const refreshButtonClick = () => {
     // 추후 새로고침 기능 추가
   };
@@ -27,13 +12,6 @@ export function ReservationCallPage() {
     <>
       <div className='flex items-center justify-center'>
         <div className='flex w-full flex-col'>
-          <ChangeToggle
-            isOpen={isOpen}
-            onToggle={toggleCallList}
-            selectedTab={selectedTab}
-            onSelect={selectTab}
-          />
-
           <div className='space-y-[1.5rem]'>
             {callListData.map((call, index) => (
               <CallList
